refactor: 타입 정의 파일 중앙화 및 디렉토리 구조 개선 (src/types)
