+++ README.md
@@ -97,7 +97,7 @@ npm run dev
 - AI 활용: Gemini로 자세한 내용 프롬프트로 작성, codex로 프로젝트 아키텍처 및 스켈레톤 코드 작성.
 - 다음 주 계획: page.tsx, layout.tsx 구현, 목 데이터 출력 해보기
 
-### Week 2 (25.12.29 - 26.01.04)
+### Week 2 (25.12.29 - 26.01.04)z`
 - 작업 내역: 
   - 더미 데이터로 페이지에 연결
   - 뷰 상태(Zustand)와 줌·패닝 동작을 정돈
@@ -154,3 +154,17 @@ npm run dev
 
     ![week4_screenshot#1.png](docs/screenshots/week4_screenshot%231.png)
 - 다음 주 계획: 완성된 백엔드와 연동하여 결과 표시 및 디버깅
+
+### Week 5 (26.01.19 ~ 26.01.25)
+- 작업 내역:
+- AI 활용:
+- 완료 기능:
+- 다음 주 계획:
+
+
+
+### Week 6 (26.01.26 ~ 26.02.01)
+- 작업 내역:
+- AI 활용:
+- 완료 기능:
+- 다음 주 계획:
