+++ src/types/genome.ts
