+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./genome";
+export * from "./analysis";
+export * from "./ui";
