+++ src/types/ui.ts
@@ -1,0 +1,2 @@
+// Shared UI-related types can be added here.
+export {};
