+++ store/useViewStore.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { create } from "zustand";
-import type { GenomeCanvasViewState } from "@/lib/types/Genome";
+import type { GenomeCanvasViewState } from "@/types";
 
 type ViewStore = {
   viewState: GenomeCanvasViewState;
